backend/requirements.txt
# --- Web Server ---
# For running the API and handling requests
fastapi
uvicorn
python-multipart


tensorflow==2.15.0
tf-keras==2.15.0

# --- Core Analysis Libraries ---
# For facial emotion recognition (replaces fer)
deepface
# For audio transcription
faster-whisper
# For eye tracking (Face Mesh)
mediapipe
# For text sentiment analysis
vaderSentiment

# --- Data & Signal Processing ---
# For numerical operations, audio, and video processing
numpy
librosa
opencv-python


sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
supabase